services/auth_service/requirements/base.txt
# services\auth_service\requirements\base.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
sqlalchemy[asyncio]
asyncpg
pydantic
bcrypt
pyjwt
redis
loguru
alembic
python-decouple
pydantic-settings
structlog
httpx
